boards: arm: qemu_cortex_m3: Refactor to use mps2_an385.
Previously qemu_cortex_m3 was emulating ti_lm3s as the underlying
soc. This emulated SOC inside the qemu doesn't have the MPU
enabled. Hence we are moving to a newer SOC which is the mps2_an385.
This patch will move the qemu_cortex_m3 from ti_lm3s to mps2_an385.
This patch will also enable the userspace.

Signed-off-by: Adithya Baglody <adithya.nagaraj.baglody@intel.com>