samples: mpu: mem_domain_apis_test: Updated the stack size.
The stack size will now be the same size as that of the
main stack. Also the creation of the threads should be done with
the actual stack size. This will be problem when MPU guard is
enabled . Which causes the stack size calculation to be offset by
MPU guard size.

Signed-off-by: Adithya Baglody <adithya.nagaraj.baglody@intel.com>